fix: resolve medicine_lookup_spec failures and improve policy consistency
